Add GetResource on HgiTexture to expose direct access to the Hgi backend's gpu resource handle as uint64_t.
This can be supported by all backends of interest, so we are at least not adding something in Hgi that couldn't be implemented by one of the backends.

It is however only exposed to make transition from GL to Hgi in Storm cleaner.
We can avoid casting to HgiGL for Hgi handles by using this new GetResource instead.

It should be noted however that eventually all the gl code in Storm needs to use Hgi and Hgi handles, because GetResource will not help us pass a Metal texture into any GL calls that remain in Storm while we transition.

(Internal change: 2075432)