Plug-in loading debug messages
- TF_DEBUG PLUG_REGISTRATION now warns about non-existent lib file

If TF_DEBUG enables reporting PLUG_REGISTRATION, a registered library
plug-in will warn if the specified library file doesn't exist
(even if the plug-in won't actually be loaded).  This makes it easy to
catch errors in plugInfo or library installation, without e.g. needing
a scene which tries to load all plug-ins.

The check adds filesystem calls (TfIsFile) but those are made only
if the TF_DEBUG PLUG_REGISTRATION is enabled.

New message:
  Unable to read library plugin 'CollectionSet' at '/.../lib64/libusd.so'.

- Debug message shows total plug-in loading time.

- More TF_DEBUG PLUG_REGISTRATION "Already registered" info

The warning:
  Already registered shared library plugin 'ShotCamera' - not registering /...so

Now also show not just the path which isn't being registered, but the
path where it was already registered.  This helps one verify that
the desired path is "winning".

Caveat: the message is quite long.

- Improved formatting of PLUG_INFO_SEARCH messages.

Aligns the paths in the "Will check" versus "Did check" messages.

- Added newlines to PLUG_REGISTRATION messages.

This helps registration messages stand out in the log; they're the
outcome of other more verbose messages in the log with TF_DEBUG="PLUG*".

The affected message is e.g.:
  Registering shared library plugin 'usdShaders' at '/.../lib64/usd/plugin/usd/usdShaders.so'

(Internal change: 2076393)