Automatically generate clip manifest layers
USD will now automatically generate a clip manifest based on
the clips specified in a clip set if one is not explicitly
provided by the user. The generated manifest is an anonymous
layer that contains an attribute spec for each attribute
found in the clip layers that have time samples.

Usd_ClipCache's lifeboat data is now managed by Usd_ClipCache
itself. This allows the cache to reuse generated manifests
during change processing. Calling UsdStage::Reload will
cause all generated manifests to be regenerated. This allows
users to force an update when new attributes are authored
into clip layers.

UsdClipsAPI::GenerateClipManifest has been added to make
creating a clip manifest more convenient. Users can call
this to create a manifest and export it to be reused on
subsequent runs, avoiding the cost of auto-generation.

This change is being done in preparation for changes that
will make value resolution depend on the clip manifest to
determine whether an attribute has a value in the associated
clip set.

(Internal change: 2076844)