Change SdrShaderNode's URI API to better reflect the reality that a node's definition may be provided by a different source/asset than its implementation.
GetSourceURI() and GetResolvedSourceURI() are being replaced by
GetResolvedDefinitionURI() and GetResolvedImplementationURI().  This
distinction is required to properly represent RenderMan C++ shaders
properly, because the .args file _defines_ the shader, while it is a
.so file that provides the implementation.

Note that the two URI's found in NdrDiscoveryResult (uri and resolvedUri)
*do not change*, because Discovery only deals with definitions, and it
is actuallly important to retain both the "raw" discovered uri and the
resolved uri, so that parsers can correctly use ArResolver::FetchToLocalResolvedPath()
prior to accessing the data.  It is therefore entirely the parser's responsibility
to compute the implementation URI and pass it on to the SdrShaderNode
it creates.

Note 2: I also removed the test of SdrShaderNode::GetInfoString() from
shaderParserTestUtils.py, since it hardcodes expected user info string output,
which is pretty brittle, annd probably unnecessary to test?

Also remove unused NdrRegistry::GetNodeByURI() until we decide how
to account for subIdentifiers; as it is, definitions from MaterialX, MDL, or
UsdShade discovery will produce undefined behavior, since many nodes
will have the same definition URI.

(Internal change: 2078091)