Fixes a crash when reparenting an instance root
When an instance root contains no imageable instances by Hydra (e.g.,
does not contain any rprims) we would lose track of that instancer and
subsequent edits to it result in undefined behavior.

This change makes it so that we always track all instancers regardless
of what they contain, so that we can correctly update them when edits
to them come in.

Fixes #1245

(Internal change: 2078294)
(Internal change: 2078865)
(Internal change: 2078882)