Move addition of time samples at value clip boundaries into Usd_Clip
Value clips introduce time samples at their boundaries and
at each time specified in the 'times' metadata for clip sets.
Previously, the former was handled in UsdStage, while the
latter was handled in Usd_Clip. This change centralizes this
behavior in Usd_Clip. This will be needed for upcoming changes
to clip interpolation. As a bonus, this lets us remove a good
chunk of code in UsdStage that was essentially doing the same
thing as code in Usd_Clip.

(Internal change: 2078331)