Add time sample for first value clip's start time
Time samples are meant to be added at all clip boundaries
and all entries in the 'times' metadata for clip sets.
However, we failed to add a time sample for the start
time of the very first clip since internally its start
time was represented as -inf. We now keep track of the
authored start time for the clip and use that when
dealing with time samples.

This bug could lead to incorrect time sample behavior
in certain cases. See the new test case in testUsdValueClips
(test_MultipleClipsWithSomeTimeSamples3) for an example.

(Internal change: 2078341)