Add ability to interpolate missing values for clips from surrounding clips
This feature lets value resolution interpolate values from
surrounding clips when it encounters a clip with no authored
time samples for an attribute declared in the manifest with
no default value. This is off by default but can be enabled
for a clip set by setting 'interpolateMissingClipValues = true'
in the clip set definition -- see new methods on UsdClipsAPI.

For example, if we have 3 clips and the 2nd clip has no
authored time sample and no default value is specified in
the manifest, the default behavior is to introduce a 'fake'
time sample at the start time of the 2nd clip with the
fallback value for the attribute's type. With this feature
enabled, this fake time sample will no longer be created.
Instead, when a time sample is queried from the 2nd clip,
value resolution will interpolate a value from the authored
time samples in the 1st and 3rd clip.

This provides better results for sparsely-authored clips
but requires value resolution to search for clips with
authored time samples. In the worst case, this means
opening and querying every value clip in the clip set.
Users can avoid this cost by authoring value blocks in
the manifest at the activation time of each clip that
does *not* contain authored samples. Value resolution
will use this information to guide its search instead of
opening each clip.

(Internal change: 2078841)