Move UsdGeomPrimvar::IsPrimvarRelatedPropertyName to UsdGeomPrimvars::CanContainPropertyName.
This move/rename is being done to be consistent with naming for similar APIs which check for propery name prefixes.
This renaming will also help when we have APIs on all prims on the lines of PropertyNameBelongsToSchema, which will associate a propertyName with a schema (Types or Applied).

(Internal change: 2082167)
(Internal change: 2082819)