[build_usd.py] on macOS include ABI flags in python library name
Prior to this change build_usd.py assumed the name of the python
library on MacOS is always "libpython#.#.dylib".  However, in practice
the library name may also contain ABI flags that indicate how this
python installation was compiled.

This patch is relevant for Python 2.7 and Python 3.x builds up to
3.7. Starting with Python 3.8 there is no longer an ABI distinction
when the interpreter is compiled with pydebug ("d") or
pymalloc ("m"). Starting with Python 3.3 and newer there is no longer
an ABI distinction for wide Unicode ("u") support.