[HdSt] Speed up wireframe rendering!
The old wireframe algorithm was to render 3x: base color, selection highlight, rollover highlight.  For the second two, we'd discard faces which weren't selection/rollover highlighted.  This works around an issue where, if you render them all as part of the same mesh, if an edge has a selected face and an unselected face adjacent, the rasterization order of those two faces is random so you randomly get the selection color or not on the edge.

The new approach is to render the wireframe mesh once, but in the GS, offset faces with selection applied.  This is ok since we're already using the GS anyway to come up with edge distance.  We use a small eps, multiplied by a number which is different for any combination of selection modes, to order faces by selection state.  This prevents Z-fighting between adjacent faces.

(Internal change: 2090159)