Remove 'realPath' parameter from SdfLayer::CreateNew and New
This parameter allowed newly-created layers to be saved to a
different location than the resolved identifier. This was a
remnant from the Pixar codebase and is no longer being used.
Removing this simplifies the API and makes it so C++ clients
don't have to pass the empty string to get to the next argument.

While here, also cleaned up some outdated documentation.

(Internal change: 2090367)