Share HgiCmds objects across buffer and compute work. Convert copyComputation to Hgi. Convert vboMemoryManager to Hgi. Convert vboSimpleMemoryManager to Hgi. Convert interleavedMemoryManager to Hgi.
A HgiCmds object is essentially a command buffer and we don't want to create to many of those per frame.
We can, however, submit the work after a certain amount of work has been recorded to make sure the GPU isn't idle for too long.

We submit recorded work in a few places:
- during the resource registry Commit phase to submit buffer, texture modifications and compute work.
- after PrepareDraw to submit dispatch buffer work
- before calling into OpenSubdiv.

We additionally submit Hgi work in extCompGpuComputation and GPUCountVisibleInstances, but this is temporary until those pieces have been converted to Hgi.

Contribution: Jason Fielder

(Internal change: 2090660)