Deprecating UsdRelationship::BlockTargets and UsdAttribute::BlockConnections and replacing all calls to BlockTargets with SetTargets({})  and BlockConnections with SetConnections({})
The reasoning for deprecation:
Attributes in USD can be robustly blocked, as we record a special token in a Layer that indicates a
value block (SdfValueBlock).  UsdAttribute::Get() will return false for any custom attribute that has
been blocked (Get() returns true for blocked builtins that have a fallback, as blocking is defined to
return the fallback, if any).

UsdAttribute::BlockConnections() and UsdRelationship::BlockTargets() claimed to block their target
values, but in fact only author an explicit empty list.  Thus, UsdAttribute::GetConnections() and
UsdRelationship::GetTargets() will return true for any "blocked" connection/relationship, because
value resolution resolved an authored, empty list (which might be empty because of listOp deletions
as well as explicit-empty).

In a usda file you can set a rel's value to None, which seems like a block, but the parser just treats
that as syntactic sugar for "explicit empty list" - no crate file can contain a "None" value for a listOp,
and we think introducing the concept of "blocked" into listOps might be more work/disruption than
it's worth.

So we're deprecating the BlockConnections and BlockTargets function in favor of calling
SetConnections and SetTargets with an empty list so that users will understand that they're not
actually able to author a true "block".

(Internal change: 2090905)