arch: Speed up first assets read of the day by omitting atime update
- What is the "atime" meta data ?
According to Redhat document, "atime" is "the last time the file (or
directory) data was accessed". atime is updated when the file is read.
This means writing to disk occurs even when just reading a file.
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/6/html/global_file_system_2/s1-manage-atimeconf

- What is the "relatime" mount option ?
To avoid this issue, the Linux has "relatime" mount option. By this
mount option, the Linux only update atime per one day (default). And
the mount option is defaultly enabled on modern Linux distribution (for
example Ubuntu 20.04).
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/6/html/power_management_guide/relatime
The mount option of a device could be checked by this command on Linux.
    $ mount | grep relatime

- What is a issue of "relatime" mount option ?
Despite this feature, the atime update will be performed on the first
read of the day still.

- How much time does "atime update" spend on USD ?
The atime update spends about 11.4 micro seconds per a file. This time
is measured with ArchPRead().

atime update enabled: 13.908 (micro seconds)
atime update disabled: 2.494 (micro seconds)
Average of ten times measurement. The time is Linux monotonic time.

So if a USD file has 100,000 USD file references, atime update causes
1.14 second delay.

- What this patch do ?
This patch avoids this issue by omitting atime update.

Signed-off-by: Masashi Honma <masashi.honma@gmail.com>