UsdShadeConnectableAPIBehavior allows plugins to customize connectability by prim type.
I started with the UsdGeomBoundableComputeExtent idea of a
function callback registry.  In the case of connectability we
currently need two callbacks, so I first bundled them as a struct,
but it started to feel like a raw form of vtbl, and I switched
to feeling it is maybe cleaner or less surprising to use an
actual UsdShadeConnectableBehavior class which plugins can
derive from.  This also allows the possibility of having some
"base" behavior that can be shared.

Using C++ OOP mechanisms like this requires allocating an instance
of the behavior object.  The registry is long-lived and can own
the instance, but it also caches lookups across the prim type
hierarchy which can lead to sharing of the same behavior.
I ended up just using a std::shared_ptr to share entries, although
arguably this is static (compile-time) behavior and should not
need to pay any runtime reference accounting.

UsdShadeConnectableAPI now consults its behavior registry
where appropriate.  I split the behavior as a separate file
/ class since most clients do not need to know about this detail.

(Internal change: 2092334)
(Internal change: 2092481)