Convert domelight to use Hgi compute and texture views.
The old code didn't explicitly bind a sampler, and relied on the default sampler state set up by HgiGLTexture. However there was an HgiSampler object created already for the source texture, but it's parameters were wrong. So I've moved the code over to use the input resource's sampler object, and fixed up the properties of it to clamp and linear mip filter.

Contribution: Jason Fielder

(Internal change: 2093612)