Add compute 'queues' so that we can shared HgiComputeCmds as much as possible between computations.
HgiComputeCmds represents a command buffer in Metal and Vulkan and we do not want to create many of those.
So sharing of command buffers is the primary motivation of this change.

We can also reduce the number of barriers we need to set by grouping computations in the correct queue.
That way we only have to do synchronization (barriers) between queues submissions.

I don't expect to see much performance difference with the removal of the barriers because our computations are executed serial on the CPU. We appear to be mostly CPU bound in our unit tests.

By using a hardcoded (but expandable) set of 'compute queues' we avoid requiring a more complicated dependency map between computations and we avoid having to 'sort' computations on the CPU before executing them. It does require the programmer to generally understand what 'compute queue' is suitable to put a computation into.

This seems ok for the current set of computations since they aren't too complex.
If more complex chaining of computations is needed, it might be useful to build a smarter system around the queueing of computations.

(Internal change: 2096910)