sdf: add hints about layer contents
Add a layer hints member that provides information about layer
contents.  The first hint (maybeHasRelocates) indicates that there
might be relocates in a layer.  If maybeHasRelocates is false, the
layer contains no relocates.  This hint is used by Pcp to skip layers
that do not contain relocates.  Hinting is conservative; a value of
maybeHasRelocates = true does not guarantee that any relocates fields
exist in the layer.  Authoring to the layer will reset its hints to a
correct but sub-optimal value.

(Internal change: 2096913)