UsdInagingGLEngine::TestIntersection should also call populate
Its possible TestIntersection could be called before any call to
UsdImagingGL::Render is made which was the only place prims were
populated. This changes makes it so that either a Render or
TestIntersection call will populate the prims via call to PrepareBatch.

Note that with the addition of PrepareBatch to TestIntersection any call to TestIntersection should add frame (timeCode) parameter to renderparams.

(Internal change: 2097639)