Speed up Clear() call on resolved attribute caches.
This is a fix to a performance regression that was detected when we
removed 'purpose' from the value cache. The issue is that we now
populate the purpose cache at every update, and we clear it at
every update (there's an opportunity here to do better). This new
clearing is quite expensive in very heavy scenes. This fix generally
improves the cost of clearing caches. Other than this issue, we haven't
noticed any other performance degradataion due to not having 'purpose'
in the value cache.

(Internal change: 2098408)