[Glf,HdSt] Cleaned-up glf lighting resource binding
Removed the need for the HdSt resource binder to
know how many uniform blocks and textures will
be used by the Glf simple lighting context.

This adds a little API to GlfBindingMap, which we
plan to obsolete, but this removes one more
fragile dependency between Glf and HdSt.

(Internal change: 2098548)