Support multiple command buffers in Metal to handle the Storm use case where we may have a blitCmds and a computeCmds recording at the same time.
We try to minimize the creation of additional command buffers in Metal for efficiency as they are heavy weight object. Encoders are lightweight, but we can only have one active encoder per command buffer. So we cannot have a blitCmds encoder and a computeCmds encoder recording at the same time unless we create an additional command buffer.

(Internal change: 2099174)