Hgi guarantees GPU-GPU synch between SubmitCmds calls.
In Vulkan this would mean inserting a semaphore between two command buffers.
In Metal, the sync is implicitely guaranteed by the Metal driver when command buffers are submitted one after the other.
In HgiGL we insert ALL_BARRIERS to ensure all GPU writes are visible to any commands in the next SubmitCmds calls.

Since HgiGL does not know what barriers are appropriate (it only knows it is supposed to provide synchronization between SubmitCmds) we use ALL_BARRIERS_BIT.
We could tweak this later if we see performance issues by picking only the BITS we know Storm needs.

This is a change required for the next change that switched GPU frustum culling to Hgi.
We would need to add additional barriers: GL_COMMAND_BARRIER_BIT and GL_UNIFORM_BARRIER_BIT, but ALL_BARRIERS seems more flexible on the Hgi level, as long as it does not cause perf regressions.

(Internal change: 2099677)